v2/iris-azure-ml/api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
joblib==1.3.2
numpy==1.24.3
pydantic==2.5.0
scikit-learn==1.3.2
python-dotenv==1.0.0